Decided on a flow of operations:
1) Build serialized documents
2) Incorporate serialized document statistics into a bare model
3) Serialize bare model
4) Specify any term reduction or configuration settings and serialize resulting model
5) Generate normalized document vectors
6) Serialize configured model
7) Cluster

Added preliminary support for the flow of operations as described above.
Fixed PDDP to work as before in clusterbase.
Added support for directory traversal in building phase 1 and phase 2 serialized documents.
Added lots of code to BirchClusterOptions and preliminary support for it in BirchKmeans.
Fixed some naming bugs in serialized document generation.