Added toString method to BirchCluster, so we can easily print out results when clustering is complete.
Changed type of maxTrialsForReasonableEffort in BirchClusterOptions to be a double, this field will represent a floating point value between 0 and 1 that is a percentage of all clusters to check for best effort approaches.
Added methods to BirchKmeans to set/get a BirchClusterOptions object.

Changed clustering logic so that useGlobalDictionaryAndBuildNormalizedVectors requires a clusteringOptions object to be set prior to being called.  This is because no changes can occur to the model once this method is invoked. And since most configuration methods are now private, a BirchClusterOptions object is the only way to configure a BirchKmeans object.

Added support for different clustering approaches to BirchKmeans: BEST_FIT_ALLOCATION and REASONABLE EFFORT (greedy was already in place).