Removed extra newline from BirchClusterOptions toString().
Added shorter prototype for createSerializedDocuments to BirchDocumentFactory

BirchKmeans no longer throws an exception from the constructor
Fixed the way we turn bp1 to bp2 in buildGlobalDictionaryFromSerializedRawDocuments
deserializeBirchKmeans can throw a generic exception or a FileNotFoundException this is used to help the driver know what to do
Added a fix to useNHighVarianceTerms so that if the term reduction value was > total distinct terms, we don't have a fatal exception

ClusteringModel no longer throws an exception from the constructor ever

Have a working driver based on Getopt every feature is now configurable.  This required massive new code.